+++ tests/metadata/test_metadata_two_point_sacc.py
@@ -530,7 +530,7 @@ def test_extract_all_photoz_bin_combinations_reals(sacc_galaxy_xis):
 
         bin_comb = all_bin_combs[tracer_names_list.index(tracer_names_type)]
         two_point_xis = TwoPointReal(
-            XY=bin_comb, thetas=np.linspace(0.0, 2.0 * np.pi, 20)
+            XY=bin_comb, thetas=np.linspace(0.0, 2.0 * np.pi, 20, dtype=np.float64)
         )
         assert two_point_xis.get_sacc_name() == tracer_names_type[1]
 
@@ -579,7 +579,7 @@ def test_make_harmonics(sacc_galaxy_cells):
 
 def test_make_reals(sacc_galaxy_xis):
     sacc_data, _, _ = sacc_galaxy_xis
-    thetas = np.linspace(0.0, 2.0 * np.pi, 20)
+    thetas = np.linspace(0.0, 2.0 * np.pi, 20, dtype=np.float64)
 
     all_bin_combs = extract_all_photoz_bin_combinations(sacc_data)
